backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.9.1
openai==1.76.2
pypdf==5.4.0
python-docx==1.1.2
python-dotenv==1.1.0
